Added solutions to 4clojure problem 24, 25, 26, 27.
Solutions are added for finding the reverse of a sequence, finding all
the odd numbers in a sequence, to sum a sequence and for finding the
fibonacci sequence.

Test cases for each of the problems have also been added.

Note:- For finding the odd numbers there are two functions: one that
uses odd? function and one without it.